+++ $notes/changelog.md
@@ -5,6 +5,7 @@
 - Botany Pots Tiers
 - chisels and bits
 - Easy Mob Farm
+- Cuisine Delight
 # 3.17:
 - Slowed down botany pot crop growth to a bit slower than how fast they grow on the ground.
 - Removed from Qol recipes that are already added by Quark